td2/requirements.txt
asgiref==3.3.1
astroid==2.5.2
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coverage==5.5
Django==3.1.7
django-autoslug==1.9.8
django-csp==3.7
django-feature-policy==4.0.0
django-permissions-policy==4.0.0
django-tinymce==3.3.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pylint==2.7.4
pylint-django==2.4.2
pylint-plugin-utils==0.6
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
